Reformat to FLTK coding style, fix one obvious issue
The only effective code change is to disable one if(..) block where
'grp->init_sizes()' was called which must not be done inside the FLTK
library code. Only user code should call this.

Another minor change is to not include <typeinfo> and use as_group()
and as_window() where needed (but that is only in disabled code).
This is work in progress...